API1/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.6
django-filter==24.3
django-prometheus==2.3.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
idna==3.10
kafka-python==2.0.3
kombu==5.4.2
packaging==24.2
pillow==11.1.0
prometheus_client==0.21.1
promise==2.3
prompt_toolkit==3.0.50
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==11.5.0
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
